add LoggingDelegatingHandler, deprecate Log4NetDelegatingHandler
